requirements.txt
tkinter
time
requests
mysql.connector
geopy.geocoders
timezonefinder
datetime
pytz
PIL